Added basic tests for Porter::importAsync() and importOneAsync().
