+++ composer.json
@@ -19,7 +19,7 @@
   },
   "require-dev": {
     "phpunit/phpunit": "^7.1.3",
-    "mockery/mockery": "^1"
+    "bilge/mockery": "dev-issue-871"
   },
   "suggest" : {
     "connectors/http": "Provides an HTTP connector for Porter providers.",
