+++ test/MockFactory.php
@@ -1,10 +1,15 @@
 <?php
 namespace ScriptFUSIONTest;
 
+use Amp\Delayed;
+use Amp\Producer;
 use Mockery\MockInterface;
+use ScriptFUSION\Porter\Connector\AsyncConnector;
 use ScriptFUSION\Porter\Connector\Connector;
 use ScriptFUSION\Porter\Connector\ImportConnector;
+use ScriptFUSION\Porter\Provider\AsyncProvider;
 use ScriptFUSION\Porter\Provider\Provider;
+use ScriptFUSION\Porter\Provider\Resource\AsyncResource;
 use ScriptFUSION\Porter\Provider\Resource\ProviderResource;
 use ScriptFUSION\StaticClass;
 
@@ -13,11 +18,11 @@ final class MockFactory
     use StaticClass;
 
     /**
-     * @return Provider|MockInterface
+     * @return Provider|AsyncProvider|MockInterface
      */
     public static function mockProvider()
     {
-        return \Mockery::namedMock(uniqid(Provider::class, false), Provider::class)
+        return \Mockery::namedMock(uniqid(Provider::class, false), Provider::class, AsyncProvider::class)
             ->shouldReceive('getConnector')
                 ->andReturn(
                     \Mockery::mock(Connector::class)
@@ -27,26 +32,37 @@ final class MockFactory
                         ->byDefault()
                 )
                 ->byDefault()
+            ->shouldReceive('getAsyncConnector')
+                ->andReturn(
+                    \Mockery::mock(AsyncConnector::class)
+                        ->shouldReceive('fetchAsync')
+                        ->andReturn(new Delayed(0, 'foo'))
+                        ->getMock()
+                )
             ->getMock()
         ;
     }
 
     /**
-     * @param Provider $provider
-     * @param \Iterator $return
-     *
-     * @return ProviderResource|MockInterface
+     * @return ProviderResource|AsyncResource|MockInterface
      */
     public static function mockResource(Provider $provider, \Iterator $return = null)
     {
-        $resource = \Mockery::mock(ProviderResource::class)
+        $resource = \Mockery::mock(ProviderResource::class, AsyncResource::class)
             ->shouldReceive('getProviderClassName')
-                ->andReturn(get_class($provider))
+                ->andReturn(\get_class($provider))
             ->shouldReceive('fetch')
-                ->andReturnUsing(function (ImportConnector $connector) {
+                ->andReturnUsing(static function (ImportConnector $connector) {
                     return new \ArrayIterator([[$connector->fetch('foo')]]);
                 })
                 ->byDefault()
+            ->shouldReceive('fetchAsync')
+                ->andReturnUsing(static function (ImportConnector $connector) {
+                    return new Producer(static function (\Closure $emit) use ($connector) {
+                        $emit([yield $connector->fetchAsync('foo')]);
+                    });
+                })
+
             ->getMock()
         ;
